// src/api/generated/schemas/listDevicesParams.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * NervesCloud API
 * The NervesCloud API gives users full access to their
Orgs, Products, and corresponding Device fleets.

The API can be used to integrate with your own systems, providing full access to your Product and Device data.

The API is documented using the OpenAPI 3.0 specification.

 * OpenAPI spec version: 2.0.0
 */
import type { ListDevicesFiltersAlarmStatus } from './listDevicesFiltersAlarmStatus';
import type { ListDevicesFiltersConnection } from './listDevicesFiltersConnection';
import type { ListDevicesFiltersDisplayDeleted } from './listDevicesFiltersDisplayDeleted';
import type { ListDevicesFiltersHasNoTags } from './listDevicesFiltersHasNoTags';
import type { ListDevicesFiltersHealthStatus } from './listDevicesFiltersHealthStatus';
import type { ListDevicesFiltersOnlyUpdating } from './listDevicesFiltersOnlyUpdating';
import type { ListDevicesFiltersUpdates } from './listDevicesFiltersUpdates';

export type ListDevicesParams = {
/**
 * Sort By
 */
sort?: string;
/**
 * Sort Direction
 */
sort_direction?: string;
/**
 * Pagination
 */
pagination?: {
  page?: number;
  page_size?: number;
};
/**
 * Filtering
 */
filters?: {
  alarm?: string;
  alarm_status?: ListDevicesFiltersAlarmStatus;
  connection?: ListDevicesFiltersConnection;
  deployment_id?: string;
  display_deleted?: ListDevicesFiltersDisplayDeleted;
  firmware_version?: string;
  has_no_tags?: ListDevicesFiltersHasNoTags;
  health_status?: ListDevicesFiltersHealthStatus;
  identifier?: string;
  only_updating?: ListDevicesFiltersOnlyUpdating;
  platform?: string;
  search?: string;
  tags?: string;
  updates?: ListDevicesFiltersUpdates;
};
};